+++ lib/testing.dart
@@ -103,7 +103,7 @@ class InMemorySpanExporter implements SpanExporter {
   @override
   Future<void> export(List<Span> spans) async {
     if (_isShutdown) {
-      throw StateError('InMemorySpanExporter is shutdown');
+      return;
     }
     _spans.addAll(spans);
   }