+++ test/testing_utils/in_memory_span_exporter.dart
@@ -21,7 +21,7 @@ class InMemorySpanExporter implements SpanExporter {
   @override
   Future<void> export(List<Span> spans) async {
     if (_isShutdown) {
-      throw StateError('Exporter is shutdown');
+      return;
     }
     _spans.addAll(spans);
   }